fix a bug (when plural.c does not exist)

+++ build.sh
@@ -128,7 +128,9 @@ if $do_build; then
 	if [ $mkst != 0 ]; then
 	    echo
 	    echo '===== workaround ===='
-	    cp $SRCDIR/intl/plural.c $SRCDIR/intl/plural.c.NG
+	    if [ -f $SRCDIR/intl/plural.c ]; then
+		cp $SRCDIR/intl/plural.c $SRCDIR/intl/plural.c.NG
+	    fi
 	    cp $SRCDIR/intl/plural.c.OK $SRCDIR/intl/plural.c
 	    echo '===== try again ===='
 	    make -j ${BUILD_PARALLELISM} ${opt_mflags}
